wedive-web/scripts/requirements.txt
google-genai>=0.2.0
jinja2
pandas
requests
firebase-admin
